--- src/main/java/xyz/tomorrowlearncamp/bookking/domain/order/entity/enums/OrderStatus.java
@@ -1,6 +1,0 @@
-package xyz.tomorrowlearncamp.bookking.domain.order.entity.enums;
-
-public enum OrderStatus {
-    COMPLETED,
-    CANCELLED
-}
