+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/order/enums/OrderStatus.java
@@ -1,0 +1,6 @@
+package xyz.tomorrowlearncamp.bookking.domain.order.enums;
+
+public enum OrderStatus {
+    COMPLETED,
+    CANCELLED
+}
